requirements.txt
aiogram>=3.3,<4.0
python-dotenv>=1.0
asyncpg>=0.29
fastapi>=0.111
uvicorn[standard]>=0.30
qrcode[pil]>=7.4
aiogram>=3.4
asyncpg>=0.29
python-dotenv>=1.0
uvloop>=0.19; sys_platform == "linux"
